add scan qr code to homescreen when other servers are present
